requirements.txt
# Core dependencies for Latency Spike Investigator
streamlit>=1.28.0
requests>=2.31.0
redis>=5.0.0
google-generativeai>=0.3.0

# Additional dependencies for data handling and visualization
pandas>=2.1.0
plotly>=5.17.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Health check endpoint (optional)
flask>=2.3.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0